+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 
 ## [2.4.0] - 2023-06-14
+- Upgrade python-lib to v2.10.1
 - Replace arg_parser.arg_parse2 with gen_class.ArgParser.
 
 ### Changed